+++ CHANGELOG.md
@@ -4,7 +4,16 @@
 
 ### Added
 
-- Two factor authentication - Support method TOTP
+- Module - follow-customer-trial attribute added. This attribute will allow
+  vendor to request module trial period to be same as nuvla trial period
+
+## [5.24.0] - 2022-03-07
+
+### Added
+
+- Nuvlabox - Use resource log resource
+- Deployment - Use resource log resource
+- Resource Log - new `resource-log` for recording logs
 
 ### Changed
 
@@ -12,6 +21,7 @@
 - Cleaned up dead code server
   start [#648](https://github.com/nuvla/api-server/issues/648)
 - Two factor authentication - Bigger email token (6 digits)
+- Two factor authentication - Support method TOTP
 - NuvlaBox Cluster - improved cluster update workflow, with added status notes
   for clarity.
 - Kafka - Allow loading Kafka producer conditionally based on